# Conflicts:
#	pubspec.lock
#	pubspec.yaml

+++ lib/event_list/controller/event_list_controller.dart
@@ -1,14 +1,16 @@
+import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:where_hearts_meet/create_event/model/event_response_model.dart';
 
 import '../../utils/consts/screen_const.dart';
 import '../../utils/controller/base_controller.dart';
+import '../../utils/dialogs/confirmation_dialog.dart';
 import '../../utils/dialogs/pop_up_dialogs.dart';
 import '../service/event_list_service.dart';
 
 class EventListController extends BaseController {
   final _eventListService = EventListService();
-  List<EventResponseModel> eventsList = [];
+  List<EventResponseModel>? eventsList;
   late EventsCreated eventsCreated;
   late String pageTitle;
   late bool forSelf;
@@ -29,32 +31,40 @@ class EventListController extends BaseController {
   }
 
   Future<void> eventsListCreatedByUser() async {
-    setBusy(true);
     final response = await _eventListService.eventsListCreatedByUserApi();
-    if (response.isNotEmpty) {
+    if (response != null && response.isNotEmpty) {
       eventsList = response;
+    } else {
+      eventsList = [];
     }
-    setBusy(false);
     update();
   }
 
   Future<void> eventsListCreatedForUser() async {
-    setBusy(true);
     final response = await _eventListService.eventsListCreatedForUserApi();
-    if (response.isNotEmpty) {
+    if (response != null && response.isNotEmpty) {
       eventsList = response;
+    } else {
+      eventsList = [];
     }
-    setBusy(false);
     update();
   }
 
   void deleteEvent({required String eventId}) async {
-    showLoaderDialog(context: Get.context!);
-    final res = await _eventListService.deleteEventApi(eventId: eventId);
-    cancelDialog();
-    if (res != null) {
-      eventsList.removeWhere((element) => element.eventid == eventId);
-    }
-    update();
+    showConfirmationBottomSheet(
+        context: Get.context!,
+        onTapYes: () async {
+          Get.back();
+          showLoaderDialog(context: Get.context!);
+          final res = await _eventListService.deleteEventApi(eventId: eventId);
+          cancelDialog();
+          if (res != null) {
+            eventsList?.removeWhere((element) => element.eventid == eventId);
+          }
+          update();
+        },
+        text: 'Delete Event?',
+        description: 'Are you sure to delete this event?',
+        iconData: Icons.delete);
   }
 }
